Re-split threshold-driven modules by domain cohesion; document the policy (#100)
Adds a "Module splitting policy" section to docs/quality-harness-phase2.md:
splits must follow domain boundaries, never metric ceilings, and a
metric warning is a signal to find the seam, not the seam itself.

Removes the one docstring in the tree that cited a quality-harness
threshold as its reason for existing (LiveBackendClipNotesMixin),
replacing it with a domain description.

Re-splits the two worst offenders along the seams the audit found:

- live_backend_parts/browser_read.py (1022 lines, 25 warnings) mixed two
  unrelated domains: loading an instrument/effect/clip browser item into
  a track (MIDI clip .alc parsing, temp-track orchestration, note/length/
  groove import -- ~900 lines) and reading the browser catalog (tree/
  item/category lookups, ~90 lines, never mutates the Live Set). Split
  into browser_load.py (LiveBackendBrowserLoadMixin) and a trimmed
  browser_read.py (LiveBackendBrowserReadMixin).
- command_backend_handlers_tracks_clips.py (911 lines, 11 warnings)
  mixed session-clip handlers with a ~270-line block of arrangement-clip
  handlers that already has its own backend mixin
  (LiveBackendScenesArrangementMixin) but no matching handler module.
  Extracted into command_backend_handlers_arrangement.py
  (ARRANGEMENT_HANDLERS), leaving tracks_clips_handlers at 625 lines.

Pure moves, no behavior or signature changes; all 913 tests pass
unchanged plus new import-location guard tests per split. Quality
harness cycles/layer-violations stay at 0. The total warning count
(374) is unchanged by the split itself: auditing the violation metrics
shows estimated_tokens/complexity/nesting/args -- the overwhelming
majority -- are per-function properties a file split cannot fix (only
god_class_risk and gross file size respond to moving code out of a
class); reducing those requires simplifying the functions themselves,
which is out of scope for a pure-move chore. Documented as a policy
rule so this isn't rediscovered the hard way next time.
